pyproject.toml
[tool.poetry]
name = "assignment-2"
version = "0.1.0"
description = "assignment2"
authors = ["vishodhan-krishnan <krishnan.vis@northeastern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.0"
bcrypt = "^4.2.0"
openai = "^1.51.0"
streamlit = "^1.39.0"
mysql-connector-python = "^9.0.0"
pandas = "^2.2.3"
pymupdf = "^1.24.11"
env = "^0.1.0"
uvicorn = "^0.31.0"
jwt = "^1.3.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
